cri: feat: support to customize the root dir for cri plugin.
In current design, the files of cri plugin are stored in the root directory of
containerd, is same directory as other plugins, such as overlayfs.
The cri plugin synchronously presist the container's status in the file which
named `stauts.json`, and ensures consistency through a read-write locks `RWLock`.
The `ListContainers` api of cri plugin implemented getting the list of containers
and get their status one by one.

When the host having high IO pressure, those IO pressure may be produced by pulling
images or writing files, and at the same time, if the cri received container created
or exited events, that will updates the container status and the on disk checkpoint,
it will take a lot of time, and holding the `RWLock`.
The `ListContainers` will not return a response until a long time later, while having
more events need to invoke `UpdateSync` of status storage.

Make the cri plugin's root directory customizable, and allowing user can choose a
different directory for cri plugin to presist status/files of containers.

Signed-off-by: Zoe <hi@zoe.im>